moved comment class into models package for reuse in subsequent steps
